refactor: Moved logic of latest version badge to a React component
